requirements.txt
faiss-gpu
PyMuPDF
transformers
torch
